requirements.txt
# --- Core Recon ---
python-nmap>=0.7.1
dnspython>=2.6.1
netlas>=0.4.0
censys>=2.1.7
requests>=2.31.0

# --- AI & Logic ---
ollama>=0.2.1
python-dotenv>=1.0.1

# --- The "Pretty" Factor (CLI/GUI) ---
rich>=13.7.1
typer>=0.12.3

# --- Reporting ---
Markdown>=3.6
WeasyPrint>=61.2
